@@ -1,74 +1,35 @@
 {
-  "name": "frontend",
-  "version": "0.1.0",
+  "name": "found-webapp",
   "private": true,
-  "dependencies": {
-    "@react-google-maps/api": "^2.1.1",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.1",
-    "axios": "^0.21.1",
-    "classnames": "^2.2.6",
-    "cogo-toast": "^4.2.3",
-    "jwt-decode": "^3.1.2",
-    "qrcode.react": "^1.0.1",
-    "qs": "^6.9.6",
-    "react": "^17.0.1",
-    "react-cropper": "^2.1.6",
-    "react-dom": "^17.0.1",
-    "react-hook-form": "^6.15.1",
-    "react-modal": "^3.12.1",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
-    "react-select": "^4.1.0",
-    "react-tooltip": "^4.2.14",
-    "react-use": "^17.2.1",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0",
-    "sass": "^1.32.7",
-    "typescript": "^4.1.5",
-    "universal-cookie": "^4.0.4",
-    "web-vitals": "^1.1.0"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@types/node": "^20.4.9",
+    "axios": "^1.4.0",
+    "qs": "^6.11.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.14.2"
   },
   "devDependencies": {
-    "@types/axios": "^0.14.0",
-    "@types/classnames": "^2.2.11",
-    "@types/qrcode.react": "^1.0.1",
-    "@types/qs": "^6.9.5",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.1",
-    "@types/react-modal": "^3.12.0",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-select": "^4.0.13",
-    "@types/redux": "^3.6.0",
-    "@types/redux-thunk": "^2.1.0",
-    "@types/universal-cookie": "^3.0.0"
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react": "^4.0.3",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "sass": "^1.64.2",
+    "typescript": "^5.0.2",
+    "vite": "^4.4.5",
+    "vite-plugin-environment": "^1.1.3",
+    "vite-tsconfig-paths": "^4.2.0"
   }
 }
